refactor and remove unnecessary tests
